Implement namespacing from RFC-0002
