+++ crates/web5/src/apid/crypto/key_managers/mod.rs
@@ -1,0 +1,1 @@
+pub mod in_memory_key_manager;
