+++ crates/web5/src/apid/dids/data_model/mod.rs
@@ -1,0 +1,3 @@
+pub mod document;
+pub mod service;
+pub mod verification_method;
