+++ crates/web5/src/apid/dids/data_model/service.rs
@@ -1,0 +1,10 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, PartialEq, Serialize, Deserialize)]
+pub struct Service {
+    pub id: String,
+    #[serde(rename = "type")]
+    pub r#type: String,
+    #[serde(rename = "serviceEndpoint")]
+    pub service_endpoint: Vec<String>,
+}
