+++ crates/web5/src/apid/dids/data_model/verification_method.rs
@@ -1,0 +1,12 @@
+use crate::apid::crypto::jwk::Jwk;
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, PartialEq, Clone, Serialize, Deserialize)]
+pub struct VerificationMethod {
+    pub id: String,
+    #[serde(rename = "type")]
+    pub r#type: String,
+    pub controller: String,
+    #[serde(rename = "publicKeyJwk")]
+    pub public_key_jwk: Jwk,
+}
