+++ crates/web5/src/apid/dids/mod.rs
@@ -1,5 +1,6 @@
+pub mod data_model;
 pub mod methods;
+pub mod resolution;
 
 pub mod did;
 pub mod document;
-pub mod resolution_result;
