+++ crates/web5/src/apid/dids/resolution/mod.rs
@@ -1,0 +1,3 @@
+pub mod document_metadata;
+pub mod resolution_metadata;
+pub mod resolution_result;
