+++ crates/web5/src/apid/mod.rs
@@ -1,6 +1,4 @@
 pub mod credentials;
+pub mod crypto;
 pub mod dids;
 pub mod dsa;
-
-pub mod in_memory_key_manager;
-pub mod jwk;
